Finish the app: flow cards, device-derived ranges, OUI hints, lint, CI
**Setpoint range came from Homey's defaults, not the appliance.** capabilitiesOptions
was never applied, so the target-temperature slider offered values outside the unit's
18-30 °C / 0.5 °C range and a write of one would be refused. Ranges now come from the
device's own minimum/maximum/increment, and are applied on every start so devices
paired before this fix pick them up without re-pairing.

**Mode writes were not checked against the device.** AC_MODES is the union across
board families, so it necessarily contains values a given unit lacks — 'Heat' on a
cooling-only Korean model. Writing one was silently dropped, which reads as a stuck
tile. Operating mode and fan speed now validate against the device's own
supportedModes, as convenient mode and airflow direction already did.

**Flow cards.** The app had none, so it could not be automated beyond raw capability
changes. Adds triggers for an alarm being raised (with the code as a token), a filter
needing attention, and a cycle finishing, plus a condition for whether state is
arriving by push. Triggers fire on transitions, tracked against the previous value:
comparing against Homey's own copy would never fire, and firing unconditionally would
fire on every poll.

**Discovery now explains itself.** A responder that will not complete a handshake is
reported with its MAC prefix rather than dropped, so "another vendor's device" can be
told from "a Samsung appliance this app cannot talk to" — indistinguishable otherwise,
and this network has four of the former. The OUI list is used only to label, never to
decide what to probe: it cannot be complete, and skipping a host for being absent from
it would hide a real appliance. The wording keeps that distinction.

Two latent bugs found by ruff: both asyncio.create_task calls held no reference, so
the discovery job and pushed-update handlers could be garbage-collected mid-flight.
And zip() over the fan-speed parallel arrays would truncate silently if they ever
differed in length, mapping speeds to the wrong codes; mismatched lengths now yield
nothing rather than a wrong answer.

Lint is clean under a purposeful rule set rather than "ALL" — correctness and
modernisation rules earn their keep, docstring formatting would be churn. The broad
excepts stay, with the reason recorded: this app talks to two contracts it cannot
enumerate, and guessing exception types would turn recoverable reads into crashes.

Adds CI running lint, tests and the reference-coverage gate on Python 3.14, the same
interpreter Homey ships. Real changelog text replaces "Initial scaffold".

+++ lib/registry/appliances.py
@@ -17,7 +17,7 @@ don't-guess rule, so does this. Nothing that applies heat is writable.
 """
 
 from . import shared
-from .base import Registry, Spec, as_float, as_int, first_item
+from .base import Registry, Spec, as_float, as_int
 
 # --- washer ---------------------------------------------------------------
 
@@ -214,7 +214,7 @@ DEHUMIDIFIER = Registry(
              lambda rep, _r: as_float(rep.get("x.com.samsung.da.desiredHumidity")),
              lambda value, _rep: (
                  ["humidity", "vs", "0"],
-                 {"x.com.samsung.da.desiredHumidity": str(int(round(float(value))))},
+                 {"x.com.samsung.da.desiredHumidity": str(round(float(value)))},
              )),
         Spec("localthings_filter_usage", "/filter/airdustfilter/vs/0",
              shared._filter_percent),
